Include time limitation into `read_process_output`

+++ securepy/stdio.py
@@ -1,5 +1,6 @@
 import subprocess
 import sys
+import time
 import typing as t
 from functools import wraps
 from io import StringIO
@@ -213,25 +214,35 @@ class IOCage:
 
 def read_process_output(
     process: subprocess.Popen,
-    read_chunk_size: int,
-    max_size: int,
+    read_chunk_size: int = 1_000,
+    max_size: t.Optional[int] = None,
+    max_exec_time: t.Optional[t.Union[float, int]] = None
 ) -> str:
     """
-    Start reading from STDOUT, stop in case stdout limit is reached or process stops.
+    Start reading from STDOUT of given `process`, return the obtained STDOUT as str.
 
-    In case output from STDOUT will reach the max limit, the subprocess will be terminated by SIGKILL.
+    In case `max_size` is specified, process will be terminated with SIGKILL if the obtained
+    output from STDOUT will get over specified `max_size` limit (in bytes).
+
+    In case `max_exec_time` is specified, process will be terminated with SIGKILL if the
+    given process will run over that specified time (in seconds).
     """
     output_size = 0
     stdout = []
+    start = time.time()
 
     while process.poll() is None:
         stdout_chars = process.stdout.read(read_chunk_size)
         output_size += sys.getsizeof(stdout_chars)
         stdout.append(stdout_chars)
 
-        if output_size > max_size:
+        if max_size is not None and output_size > max_size:
             process.kill()
             raise MemoryOverflow(output_size, max_size, "Terminating subprocess.Popen with SIGKILL")
+
+        if max_exec_time is not None and time.time() - start > max_exec_time:
+            process.kill()
+            raise TimeoutError(f"Terminating subprocess.Popen with SIGKILL (surpassed maximum time for execution: {max_exec_time})")
 
     # Wait for process termination
     process.wait()
